aggiungto deletePaziente e gestiti i casi in cui vengono immessi valori nulli nelle form

+++ src/main/java/it/uniroma3/siw/controller/MedicoController.java
@@ -1,15 +1,16 @@
 package it.uniroma3.siw.controller;
 
 import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import it.uniroma3.siw.controller.validator.MedicoValidator;
 import it.uniroma3.siw.model.Esame;
 import it.uniroma3.siw.model.Medico;
 import it.uniroma3.siw.service.MedicoService;
@@ -19,6 +20,9 @@ public class MedicoController {
 
 	@Autowired
 	private MedicoService medicoService;
+	
+	@Autowired
+	private MedicoValidator medicoValidator;
 
 	@RequestMapping(value="/addMedico", method=RequestMethod.GET)
 	public String getAddMedicoForm(Model model) {
@@ -28,26 +32,44 @@ public class MedicoController {
 	}
 	
 	@RequestMapping(value="/addMedico", method=RequestMethod.POST)
-	public String addMedicoForm(Model model, @ModelAttribute("medico") Medico medico) {
+	public String addMedicoForm(Model model, @ModelAttribute("medico") Medico medico, BindingResult bindingResult) {
 
-		medicoService.inserisci(medico);
-		return "index";
+		this.medicoValidator.validate(medico, bindingResult);
+		
+		if(!bindingResult.hasErrors()) {
+			medicoService.inserisci(medico);
+			return "index";
+		}
+		
+		model.addAttribute("medico", new Medico());
+		
+		return "/admin/addMedicoForm";
 	}
 	
 	@RequestMapping(value="/deleteMedico", method=RequestMethod.GET)
 	public String deleteMedico(Model model,
-							   @RequestParam("medicoID") Long id,
+							   @RequestParam("medicoID") String id,
 							   @RequestParam("nome") String nome,
 							   @RequestParam("cognome") String cognome) {
 		
-		Medico medicoByNC = this.medicoService.medicoPerNomeCognome(nome, cognome);
-		Medico medicoById = this.medicoService.medicoPerId(id);
 		
-		if(medicoById.equals(medicoByNC)) {
-			this.medicoService.delete(medicoById);
+		if(!id.equals("") && !nome.equals("") && !cognome.equals("")) {
+			Medico medicoByNC = this.medicoService.medicoPerNomeCognome(nome, cognome);
+			Medico medicoById = this.medicoService.medicoPerId(Long.parseLong(id));
+			
+			if(medicoByNC == null || medicoById == null) {
+				model.addAttribute("frase", "Non è stato possibile eliminare il medico, ricontrolli i dati immessi.");
+				return "risultatiNonDisponibili";
+			}
+			
+			if(medicoById.equals(medicoByNC)) {
+				this.medicoService.delete(medicoById);
+				return "index";
+			}
+			
 		}
 		
-		return "index";
+		return "/admin/deleteMedico";
 	}
 	@RequestMapping(value="/deleteMedicoForm", method=RequestMethod.GET)
 	public String deleteMedico(Model model) {
@@ -65,11 +87,15 @@ public class MedicoController {
 	@RequestMapping(value="/esamiPerMedico", method=RequestMethod.GET)
 	public String cercaMedico(Model model, @RequestParam("nome") String nome, @RequestParam("cognome") String cognome) {
 		
-		Medico medico = this.medicoService.medicoPerNomeCognome(nome, cognome);
+		if(!nome.equals("") && !cognome.equals("")) {
+			Medico medico = this.medicoService.medicoPerNomeCognome(nome, cognome);
+			
+			List<Esame> esami = this.medicoService.getEsamiPrenotati(medico.getId());
+			model.addAttribute("esami", esami);
+
+			return "esamiPrenotati";
+		}
 		
-		List<Esame> esami = this.medicoService.getEsamiPrenotati(medico.getId());
-		model.addAttribute("esami", esami);
-		
-		return "esamiPrenotati";
+		return "/admin/cercaMedico";
 	}
 }
